requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
boto3==1.17.61
botocore==1.20.61
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
confusable-homoglyphs==3.2.0
coverage==5.5
crispy-bootstrap5==0.3.1
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.44.0
django-bootstrap-v5==1.0.2
django-comments-xtd==2.9.1
django-contrib-comments==2.1.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-login-required-middleware==0.6.1
django-mptt==0.12.0
django-storages==1.11.1
django-threadedcomments==1.2
djangorestframework==3.12.4
docutils==0.17.1
easy-thumbnails==2.7.1
flake8==3.9.1
flake8-django==1.1.1
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
